04_App/requirements.txt
dash-bootstrap-components>=1.5.0
plotly>=5.18.0
pandas>=2.0.0
databricks-sql-connector>=3.1.0
requests>=2.31.0
